Library Updated to Support Compatibility with other major Libarires
This is libarary has been upgraded by replacing the SoftwareSerial Library with that of the AltSoftSerial one. This makes it now compatibile with libraries that uese the pinchange interrupts like ENableInterrupts and also the SPI since pin 10 and 11 and no more used.

This upgrade means only the assigned pin can be used for communicating with the SIM800